refactor: change the datatype of lottoAmount from int to double
Change the datatype of lottoAmount from int to double because lottoAmount can have big number if money is big number.
To print the lottoAmount or make list, int type is needed as a parameter so change datatype in that case.

+++ src/main/java/lotto/ErrorControl.java
@@ -34,7 +34,7 @@ public class ErrorControl {
 
     public void validateNumber(String money) {
         try {
-            Integer.parseInt(money);
+            Double.parseDouble(money);
         } catch (IllegalArgumentException e) {
             throw new MyIllegalArgumentException("[ERROR] 숫자를 입력해 주세요.");
         }
